extracted core ui into separate UIModule
RootModule "AppModule" cannot be imported,
therefore a separate UIModule with the core ui is helpful
