+++ app/app.module.ts
@@ -4,23 +4,17 @@ import {FormsModule} from "@angular/forms";
 
 import {AppComponent} from './app.component';
 import {routing} from "./app.routes";
-import {LoggedInGuard} from "./session/logged-in.guard";
-import {EntityMapperService} from "./entity/entity-mapper.service";
-import {FooterComponent} from "./footer.component";
 import {AlertsModule} from "./alerts/alerts.module";
 import {NG2BootstrapModule} from "./ng2-bootstrap.module";
 import {DatabaseModule} from "./database/database.module";
 import {ConfigModule} from "./config/config.module";
-import SessionModule from "./session/session.module";
+import {SessionModule} from "./session/session.module";
 import {SyncStatusModule} from "./sync-status/sync-status.module";
 import {DashboardModule} from "./dashboard/dashboard.module";
 import {NavigationModule} from "./navigation/navigation.module";
+import {UIModule} from "./ui/ui.module";
 
 @NgModule({
-    declarations: [
-        AppComponent,
-        FooterComponent,
-    ],
     imports: [
         BrowserModule,
         routing,
@@ -32,14 +26,12 @@ import {NavigationModule} from "./navigation/navigation.module";
         SessionModule,
         SyncStatusModule,
         DashboardModule,
-        NavigationModule
+        NavigationModule,
+        UIModule
         // UserModule is lazy loaded
     ],
-    bootstrap: [AppComponent],
-    providers: [
-        LoggedInGuard,
-        EntityMapperService
-    ]
+    declarations: [AppComponent],
+    bootstrap: [AppComponent]
 })
 export class AppModule {
 }
