+++ app/ui/footer.component.ts
@@ -3,7 +3,7 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'ndb-footer',
-    templateUrl: 'app/footer.component.html'
+    templateUrl: 'app/ui/footer.component.html'
 })
 export class FooterComponent {
