+++ app/ui/ui.component.css
@@ -1,4 +1,4 @@
-ndb-app {
+ndb-ui {
     width: 100%;
     background-color: #f8f8f8;
 }
