+++ index.ts
@@ -8,3 +8,4 @@ export * from './app/entity/entity.module';
 export * from './app/session/session.module';
 export * from './app/sync-status/sync-status.module';
 export * from './app/user/user.module';
+export * from './app/ui/ui.module';
